# api/src/main/resources/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Data cryptography
  description: API that retrieves and saves encrypted data
servers:
  - url: http://localhost:8080/v1
tags:
  - name: data
paths:
  /data:
    get:
      summary: List all data
      operationId: retrieveData
      tags: [data]
      responses:
        200:
          description: The list of available data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataList"
    post:
      summary: Save and encrypts data
      operationId: saveData
      tags: [data]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewData"
      responses:
        201:
          description: The encrypted data with an id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Data"
  /data/{id}:
    put:
      summary: Update and encrypts data
      operationId: updateData
      tags: [data]
      parameters:
        - name: id
          in: path
          description: id of the data to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewData"
      responses:
        200:
          description: The encrypted updated data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Data"

  /data/{id}/decrypted:
    get:
      summary: get the decrypted value
      operationId: getDecryptedData
      tags:
        - data
      parameters:
        - name: id
          in: path
          description: id of the data to be updated
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: The decrypted data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Data"
components:
  schemas:
    Data:
      type: "object"
      required:
        - id
        - data
      properties:
        id:
          type: "integer"
          format: "int64"
        data:
          type: "string"
    NewData:
      type: 'object'
      required:
        - data
      properties:
        data:
          type: string
          maxLength: 245
    DataList:
      type: 'array'
      items:
        $ref: '#/components/schemas/Data'